# cmd/apiserver/openapi/model_channel.go
/*
 * Fledge REST API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Channel struct {

	Name string `json:"name"`

	Backend string `json:"backend"`

	Directions []Direction `json:"directions"`

	Groupby string `json:"groupby,omitempty"`

	Tag string `json:"tag,omitempty"`
}