# cmd/apiserver/openapi/model_design_schema.go
/*
 * Fledge REST API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type DesignSchema struct {

	Roles []Role `json:"roles"`

	Channels []Channel `json:"channels"`

	Tag string `json:"tag,omitempty"`
}