# cmd/apiserver/openapi/model_direction.go
/*
 * Fledge REST API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Direction struct {

	From string `json:"from"`

	To string `json:"to"`

	Tag string `json:"tag,omitempty"`
}