# cmd/apiserver/openapi/model_role.go
/*
 * Fledge REST API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Role struct {

	Name string `json:"name"`

	Tag string `json:"tag,omitempty"`
}